MdeModulePkg: Remove the encryption mask from non-leaf page table entries
The CpuPageTableLib library requires that non-leaf page table entries not
have the encryption bit set. An SEV guest always treats non-leaf page
table entries as encrypted, so it is not required to have the encryption
bit set in non-leaf entries.

Update the creation of non-leaf entries to not set the encryption bit so
that a future conversion to using the CpuPageTableLib library can be done.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>